# api-orders/docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  dto.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.OrderItemRequest'
        minItems: 1
        type: array
    required:
    - items
    type: object
  dto.CreateProductRequest:
    properties:
      description:
        example: Mouse Gamer RGB 16000 DPI
        type: string
      name:
        example: Mouse Gamer
        minLength: 3
        type: string
      price:
        example: 199.9
        type: number
      quantity:
        example: 50
        minimum: 0
        type: integer
    required:
    - description
    - name
    - price
    - quantity
    type: object
  dto.OrderItemRequest:
    properties:
      product_id:
        example: 698c0a0893c94ce530171bbb
        type: string
      quantity:
        example: 6
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.OrderItemResponse:
    properties:
      price:
        example: 199.9
        type: number
      product_id:
        example: 507f1f77bcf86cd799439011
        type: string
      product_name:
        example: Mouse Gamer
        type: string
      quantity:
        example: 2
        type: integer
    type: object
  dto.OrderResponse:
    properties:
      _id:
        example: 507f1f77bcf86cd799439011
        type: string
      created_at:
        example: "2024-02-10T12:00:00Z"
        type: string
      items:
        items:
          $ref: '#/definitions/dto.OrderItemResponse'
        type: array
      order_number:
        example: ORD-A1B2C3D4
        type: string
      status:
        example: criado
        type: string
      total:
        example: 399.8
        type: number
      updated_at:
        example: "2024-02-10T12:00:00Z"
        type: string
    type: object
  dto.ProductResponse:
    properties:
      _id:
        example: 507f1f77bcf86cd799439011
        type: string
      created_at:
        example: "2024-02-10T12:00:00Z"
        type: string
      description:
        example: Mouse Gamer RGB 16000 DPI
        type: string
      name:
        example: Mouse Gamer
        type: string
      price:
        example: 199.9
        type: number
      quantity:
        example: 50
        type: integer
      updated_at:
        example: "2024-02-10T12:00:00Z"
        type: string
    type: object
  dto.UpdateOrderStatusRequest:
    properties:
      status:
        enum:
        - criado
        - em_processamento
        - enviado
        - entregue
        example: enviado
        type: string
    required:
    - status
    type: object
  handlers.ErrorResponseDoc:
    properties:
      error:
        example: Detailed error message
        type: string
      message:
        example: User-friendly error message
        type: string
      success:
        example: false
        type: boolean
    type: object
  handlers.SuccessResponseDoc:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API backend escalável para gerenciamento de produtos e pedidos construída
    com Go, Gin Framework, MongoDB e RabbitMQ seguindo Arquitetura Hexagonal (Clean
    Architecture).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Order Management API
  version: "1.0"
paths:
  /health:
    get:
      description: Returns the health status of the API and RabbitMQ connection
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - Health
  /orders:
    post:
      consumes:
      - application/json
      description: Creates a new order with the provided items
      parameters:
      - description: Order information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order created successfully
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponseDoc'
            - properties:
                data:
                  $ref: '#/definitions/dto.OrderResponse'
              type: object
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
      summary: Create a new order
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves an order by its MongoDB ObjectID
      parameters:
      - description: Order ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponseDoc'
            - properties:
                data:
                  $ref: '#/definitions/dto.OrderResponse'
              type: object
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
      summary: Get order by ID
      tags:
      - Orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates the status of an existing order
      parameters:
      - description: Order ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order status updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponseDoc'
            - properties:
                data:
                  $ref: '#/definitions/dto.OrderResponse'
              type: object
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
      summary: Update order status
      tags:
      - Orders
  /products:
    post:
      consumes:
      - application/json
      description: Creates a new product with the provided information
      parameters:
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponseDoc'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponseDoc'
      summary: Create a new product
      tags:
      - Products
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Operações relacionadas a produtos
  name: Products
- description: Operações relacionadas a pedidos
  name: Orders
- description: Health check da aplicação
  name: Health